# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateNotificationChannelRequest:
    properties:
      channel_type:
        enum:
        - lark
        - telegram
        type: string
      config: {}
      enabled:
        type: boolean
      name:
        maxLength: 100
        type: string
    required:
    - channel_type
    - config
    - name
    type: object
...
  dto.NotificationChannelResponse:
    properties:
      channel_type:
        type: string
      config: {}
      created_at:
        type: string
      created_by:
        type: integer
      enabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateNotificationChannelRequest:
    properties:
      config: {}
      enabled:
        type: boolean
      name:
        maxLength: 100
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:10010
info:
  contact:
    name: TicketDesk
    url: https://github.com/kerbos/ticketdesk
  description: |-
    项目化工单与告警联动系统 API. 一切问题都是工单, 一切告警都必须被跟进.
    **统一响应格式**: `{code, message, data}`. 错误响应 `{code, message, details?}` (code 为字符串如 "BAD_REQUEST" / "UNAUTHORIZED").
    **速率限制**: 全站默认 300 req/min/IP, /auth/* 端点 20 req/min/IP, /webhook/* 100 req/min/IP. 超出返回 **429 Too Many Requests**, 此状态码在所有端点皆可能出现, 单独 endpoint 文档不重复声明.
  license:
    name: MIT
  termsOfService: https://github.com/kerbos/ticketdesk
  title: TicketDesk API
  version: "1.0"
paths:
...
  /api/v1/projects/{key}/notification-channels:
    get:
      description: 获取指定项目的所有通知渠道
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.NotificationChannelResponse'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取项目通知渠道列表
      tags:
      - ProjectNotificationChannel
    post:
      consumes:
      - application/json
      description: 为项目创建一个新的通知渠道
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      - description: 创建通知渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NotificationChannelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建通知渠道
      tags:
      - ProjectNotificationChannel
  /api/v1/projects/{key}/notification-channels/{id}:
    delete:
      description: 删除指定的通知渠道
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 无效的渠道 ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除通知渠道
      tags:
      - ProjectNotificationChannel
    get:
      description: 根据 ID 获取通知渠道详细信息
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NotificationChannelResponse'
              type: object
        "400":
          description: 无效的渠道 ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取通知渠道详情
      tags:
      - ProjectNotificationChannel
    put:
      consumes:
      - application/json
      description: 更新通知渠道信息
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新通知渠道请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NotificationChannelResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新通知渠道
      tags:
      - ProjectNotificationChannel
  /api/v1/projects/{key}/notification-channels/{id}/test:
    post:
      description: 发送测试消息到指定通知渠道
      parameters:
      - description: 项目 Key
        in: path
        name: key
        required: true
        type: string
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 无效的渠道 ID 或发送失败
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 测试通知渠道
      tags:
      - ProjectNotificationChannel
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer 令牌 (`Bearer <jwt>`) 或 Personal Access Token (`Bearer td_pat_xxx`)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"